refactor: EmptyState·SessionList 컴포넌트 분리
